MealRestController in memory implementation( call from Servlet) with filtration and simple user choice

+++ src/main/java/ru/javawebinar/topjava/repository/inmemory/InMemoryMealRepository.java
@@ -5,14 +5,15 @@ import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.DateTimeUtil;
 import ru.javawebinar.topjava.util.MealsUtil;
-import ru.javawebinar.topjava.util.exception.NotFoundException;
+import ru.javawebinar.topjava.web.SecurityUtil;
 
 import java.time.LocalDateTime;
-import java.time.LocalTime;
-import java.util.Collection;
+import java.util.Comparator;
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
 @Repository
@@ -21,18 +22,19 @@ public class InMemoryMealRepository implements MealRepository {
     private final AtomicInteger counter = new AtomicInteger(0);
 
     {
-        MealsUtil.meals.forEach(this::save);
+        MealsUtil.meals.forEach(meal -> save(meal, 1));
     }
 
     @Override
-    public Meal save(Meal meal) {
+    public Meal save(Meal meal, int userId) {
+        meal.setUserId(SecurityUtil.authUserId());
         if (meal.isNew()) {
             meal.setId(counter.incrementAndGet());
             repository.put(meal.getId(), meal);
             return meal;
         }
         // handle case: update, but not present in storage
-        if (repository.get(meal.getId()).getUserId() == meal.getUserId()) {
+        if (repository.containsKey(meal.getId()) && repository.get(meal.getId()).getUserId() == meal.getUserId()) {
             return repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal);
         } else {
             return null;
@@ -41,7 +43,7 @@ public class InMemoryMealRepository implements MealRepository {
 
     @Override
     public boolean delete(int userId, int id) {
-        if (repository.get(id).getUserId() == userId) {
+        if (repository.containsKey(id) && repository.get(id).getUserId() == userId) {
             return repository.remove(id) != null;
         } else {
             return false;
@@ -50,7 +52,7 @@ public class InMemoryMealRepository implements MealRepository {
 
     @Override
     public Meal get(int userId, int id) {
-        if (repository.get(id).getUserId() == userId) {
+        if (repository.containsKey(id) && repository.get(id).getUserId() == userId) {
             return repository.get(id);
         } else {
             return null;
@@ -58,17 +60,19 @@ public class InMemoryMealRepository implements MealRepository {
     }
 
     @Override
-    public Collection<Meal> getAll(int userId) {
-        return repository.values().stream()
-                .filter(meal -> meal.getUserId() == userId)
-                .sorted((m1, m2) -> m1.getDateTime().isBefore(m2.getDateTime()) ? 1 : -1)
-                .collect(Collectors.toList());
+    public List<Meal> getAll(int userId) {
+        return getAllByPredicate(meal -> meal.getUserId() == userId);
     }
 
-    public Collection<Meal> getAllFiltered(int userId, LocalDateTime startDate, LocalDateTime endDate, LocalTime startTime, LocalTime endTime) {
-        return getAll(userId).stream()
-                .filter(meal -> DateTimeUtil.isBetweenHalfOpen(meal.getDate(), startDate, endDate.plusDays(1)))
-                .filter(meal -> DateTimeUtil.isBetweenHalfOpen(meal.getTime(), startTime, endTime))
+    public List<Meal> getAllFiltered(int userId, LocalDateTime startDate, LocalDateTime endDate) {
+        return getAllByPredicate(meal -> meal.getUserId() == userId &&
+                DateTimeUtil.isBetweenHalfOpen(meal.getDateTime(), startDate, endDate.plusDays(1)));
+    }
+
+    public List<Meal> getAllByPredicate(Predicate<Meal> filter) {
+        return repository.values().stream()
+                .filter(filter)
+                .sorted(Comparator.comparing(Meal::getDateTime).reversed())
                 .collect(Collectors.toList());
     }
 }
